requirements.txt
# Core AI/ML Libraries
torch>=2.0.0
openai-whisper>=20231117
transformers>=4.30.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1
scipy>=1.10.0
numpy>=1.24.0
pygame>=2.5.0

# Text-to-Speech
edge-tts>=6.1.0
gtts>=2.4.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
ddgs>=0.7.0

# Optional - Voice Activity Detection (requires C/C++ compiler)
# If build fails on Windows, comment this line out - VAD will fallback to amplitude-based detection
# webrtcvad>=2.0.10

# Optional - Web Search API
tavily-python>=0.3.0

# Ollama Python SDK
ollama>=0.1.0

# Optional - Wake Word Detection (Porcupine)
pvporcupine>=3.0.0  # Uncomment if you want offline wake word detection
